Remove remaining `AttributeLintKind` variants
