+++ compiler/rustc_attr_parsing/src/attributes/diagnostic/check_cfg.rs
@@ -1,12 +1,10 @@
-use rustc_hir::def_id::LOCAL_CRATE;
-use rustc_middle::bug;
-use rustc_middle::ty::TyCtxt;
-use rustc_session::Session;
 use rustc_session::config::ExpectedValues;
+use rustc_session::{Session, SessionAndCrateName};
+use rustc_span::def_id::{CrateNum, LOCAL_CRATE};
 use rustc_span::edit_distance::find_best_match_for_name;
 use rustc_span::{ExpnKind, Ident, Span, Symbol, sym};
 
-use crate::lints;
+use crate::errors;
 
 const MAX_CHECK_CFG_NAMES_OR_VALUES: usize = 35;
 
@@ -64,57 +62,57 @@ fn to_check_cfg_arg(name: Ident, value: Option<Symbol>, quotes: EscapeQuotes) ->
 fn cargo_help_sub(
     sess: &Session,
     inst: &impl Fn(EscapeQuotes) -> String,
-) -> lints::UnexpectedCfgCargoHelp {
+) -> errors::UnexpectedCfgCargoHelp {
     // We don't want to suggest the `build.rs` way to expected cfgs if we are already in a
     // `build.rs`. We therefor do a best effort check (looking if the `--crate-name` is
     // `build_script_build`) to try to figure out if we are building a Cargo build script
 
     let unescaped = &inst(EscapeQuotes::No);
     if let Some("build_script_build") = sess.opts.crate_name.as_deref() {
-        lints::UnexpectedCfgCargoHelp::lint_cfg(unescaped)
+        errors::UnexpectedCfgCargoHelp::lint_cfg(unescaped)
     } else {
-        lints::UnexpectedCfgCargoHelp::lint_cfg_and_build_rs(unescaped, &inst(EscapeQuotes::Yes))
+        errors::UnexpectedCfgCargoHelp::lint_cfg_and_build_rs(unescaped, &inst(EscapeQuotes::Yes))
     }
 }
 
-fn rustc_macro_help(span: Span) -> Option<lints::UnexpectedCfgRustcMacroHelp> {
+fn rustc_macro_help(span: Span) -> Option<errors::UnexpectedCfgRustcMacroHelp> {
     let oexpn = span.ctxt().outer_expn_data();
     if let Some(def_id) = oexpn.macro_def_id
         && let ExpnKind::Macro(macro_kind, macro_name) = oexpn.kind
         && def_id.krate != LOCAL_CRATE
     {
-        Some(lints::UnexpectedCfgRustcMacroHelp { macro_kind: macro_kind.descr(), macro_name })
+        Some(errors::UnexpectedCfgRustcMacroHelp { macro_kind: macro_kind.descr(), macro_name })
     } else {
         None
     }
 }
 
 fn cargo_macro_help(
-    tcx: Option<TyCtxt<'_>>,
+    crate_name: Option<&dyn Fn(CrateNum) -> Symbol>,
     span: Span,
-) -> Option<lints::UnexpectedCfgCargoMacroHelp> {
+) -> Option<errors::UnexpectedCfgCargoMacroHelp> {
     let oexpn = span.ctxt().outer_expn_data();
-    if let Some(def_id) = oexpn.macro_def_id
-        && let ExpnKind::Macro(macro_kind, macro_name) = oexpn.kind
+    if let Some(crate_name) = crate_name
+        && let Some(def_id) = oexpn.macro_def_id
         && def_id.krate != LOCAL_CRATE
-        && let Some(tcx) = tcx
+        && let ExpnKind::Macro(macro_kind, macro_name) = oexpn.kind
     {
-        Some(lints::UnexpectedCfgCargoMacroHelp {
+        Some(errors::UnexpectedCfgCargoMacroHelp {
             macro_kind: macro_kind.descr(),
             macro_name,
-            crate_name: tcx.crate_name(def_id.krate),
+            crate_name: crate_name(def_id.krate),
         })
     } else {
         None
     }
 }
 
-pub(super) fn unexpected_cfg_name(
-    sess: &Session,
-    tcx: Option<TyCtxt<'_>>,
+pub(crate) fn unexpected_cfg_name(
+    sess_and_name: &SessionAndCrateName<'_>,
     (name, name_span): (Symbol, Span),
     value: Option<(Symbol, Span)>,
-) -> lints::UnexpectedCfgName {
+) -> errors::UnexpectedCfgName {
+    let sess = sess_and_name.sess;
     #[allow(rustc::potential_query_instability)]
     let possibilities: Vec<Symbol> = sess.psess.check_config.expecteds.keys().copied().collect();
 
@@ -149,12 +147,12 @@ pub(super) fn unexpected_cfg_name(
     }
 
     let code_sugg = if is_feature_cfg && is_from_cargo {
-        lints::unexpected_cfg_name::CodeSuggestion::DefineFeatures
+        errors::unexpected_cfg_name::CodeSuggestion::DefineFeatures
     // Suggest correct `version("..")` predicate syntax
     } else if let Some((_value, value_span)) = value
         && name == sym::version
     {
-        lints::unexpected_cfg_name::CodeSuggestion::VersionSyntax {
+        errors::unexpected_cfg_name::CodeSuggestion::VersionSyntax {
             between_name_and_value: name_span.between(value_span),
             after_value: value_span.shrink_to_hi(),
         }
@@ -169,7 +167,7 @@ pub(super) fn unexpected_cfg_name(
             .span_to_snippet(name_span)
             .map_or(true, |snippet| !snippet.contains("r#"))
     {
-        lints::unexpected_cfg_name::CodeSuggestion::BooleanLiteral {
+        errors::unexpected_cfg_name::CodeSuggestion::BooleanLiteral {
             span: name_span,
             literal: boolean,
         }
@@ -189,7 +187,7 @@ pub(super) fn unexpected_cfg_name(
                 if !possibilities.is_empty() {
                     let possibilities =
                         possibilities.iter().copied().cloned().collect::<Vec<_>>().into();
-                    Some(lints::unexpected_cfg_name::ExpectedValues { best_match, possibilities })
+                    Some(errors::unexpected_cfg_name::ExpectedValues { best_match, possibilities })
                 } else {
                     None
                 }
@@ -198,37 +196,37 @@ pub(super) fn unexpected_cfg_name(
             let best_match = Ident::new(best_match, name_span);
             if let Some((value, value_span)) = value {
                 if best_match_values.contains(&Some(value)) {
-                    lints::unexpected_cfg_name::CodeSuggestion::SimilarNameAndValue {
+                    errors::unexpected_cfg_name::CodeSuggestion::SimilarNameAndValue {
                         span: name_span,
                         code: best_match.to_string(),
                     }
                 } else if best_match_values.contains(&None) {
-                    lints::unexpected_cfg_name::CodeSuggestion::SimilarNameNoValue {
+                    errors::unexpected_cfg_name::CodeSuggestion::SimilarNameNoValue {
                         span: name_span.to(value_span),
                         code: best_match.to_string(),
                     }
                 } else if let Some(first_value) = possibilities.first() {
-                    lints::unexpected_cfg_name::CodeSuggestion::SimilarNameDifferentValues {
+                    errors::unexpected_cfg_name::CodeSuggestion::SimilarNameDifferentValues {
                         span: name_span.to(value_span),
                         code: format!("{best_match} = \"{first_value}\""),
                         expected: get_possibilities_sub(),
                     }
                 } else {
-                    lints::unexpected_cfg_name::CodeSuggestion::SimilarNameDifferentValues {
+                    errors::unexpected_cfg_name::CodeSuggestion::SimilarNameDifferentValues {
                         span: name_span.to(value_span),
                         code: best_match.to_string(),
                         expected: get_possibilities_sub(),
                     }
                 }
             } else {
-                lints::unexpected_cfg_name::CodeSuggestion::SimilarName {
+                errors::unexpected_cfg_name::CodeSuggestion::SimilarName {
                     span: name_span,
                     code: best_match.to_string(),
                     expected: get_possibilities_sub(),
                 }
             }
         } else {
-            lints::unexpected_cfg_name::CodeSuggestion::SimilarName {
+            errors::unexpected_cfg_name::CodeSuggestion::SimilarName {
                 span: name_span,
                 code: best_match.to_string(),
                 expected: None,
@@ -239,7 +237,7 @@ pub(super) fn unexpected_cfg_name(
             names_possibilities.sort();
             names_possibilities
                 .iter()
-                .map(|cfg_name| lints::unexpected_cfg_name::FoundWithSimilarValue {
+                .map(|cfg_name| errors::unexpected_cfg_name::FoundWithSimilarValue {
                     span: name_span,
                     code: format!("{cfg_name} = \"{name}\""),
                 })
@@ -253,14 +251,14 @@ pub(super) fn unexpected_cfg_name(
         let expected_names = if !possibilities.is_empty() {
             let possibilities: Vec<_> =
                 possibilities.into_iter().map(|s| Ident::new(s, name_span)).collect();
-            Some(lints::unexpected_cfg_name::ExpectedNames {
+            Some(errors::unexpected_cfg_name::ExpectedNames {
                 possibilities: possibilities.into(),
                 and_more,
             })
         } else {
             None
         };
-        lints::unexpected_cfg_name::CodeSuggestion::SimilarValues {
+        errors::unexpected_cfg_name::CodeSuggestion::SimilarValues {
             with_similar_values: similar_values,
             expected_names,
         }
@@ -276,29 +274,29 @@ pub(super) fn unexpected_cfg_name(
         } else {
             None
         };
-        lints::unexpected_cfg_name::InvocationHelp::Cargo {
+        errors::unexpected_cfg_name::InvocationHelp::Cargo {
             help,
-            macro_help: cargo_macro_help(tcx, name_span),
+            macro_help: cargo_macro_help(sess_and_name.crate_name, name_span),
         }
     } else {
-        let help = lints::UnexpectedCfgRustcHelp::new(&inst(EscapeQuotes::No));
-        lints::unexpected_cfg_name::InvocationHelp::Rustc {
+        let help = errors::UnexpectedCfgRustcHelp::new(&inst(EscapeQuotes::No));
+        errors::unexpected_cfg_name::InvocationHelp::Rustc {
             help,
             macro_help: rustc_macro_help(name_span),
         }
     };
 
-    lints::UnexpectedCfgName { code_sugg, invocation_help, name }
+    errors::UnexpectedCfgName { code_sugg, invocation_help, name }
 }
 
-pub(super) fn unexpected_cfg_value(
-    sess: &Session,
-    tcx: Option<TyCtxt<'_>>,
+pub(crate) fn unexpected_cfg_value(
+    sess_and_name: &SessionAndCrateName<'_>,
     (name, name_span): (Symbol, Span),
     value: Option<(Symbol, Span)>,
-) -> lints::UnexpectedCfgValue {
+) -> errors::UnexpectedCfgValue {
+    let sess = sess_and_name.sess;
     let Some(ExpectedValues::Some(values)) = &sess.psess.check_config.expecteds.get(&name) else {
-        bug!(
+        panic!(
             "it shouldn't be possible to have a diagnostic on a value whose name is not in values"
         );
     };
@@ -327,13 +325,13 @@ pub(super) fn unexpected_cfg_value(
             .iter()
             .take(max_suggestions)
             .copied()
-            .map(|name| lints::unexpected_cfg_value::ChangeNameSuggestion {
+            .map(|name| errors::unexpected_cfg_value::ChangeNameSuggestion {
                 span: name_span,
                 name,
                 value,
             })
             .collect::<Vec<_>>();
-        lints::unexpected_cfg_value::CodeSuggestion::ChangeName { suggestions }
+        errors::unexpected_cfg_value::CodeSuggestion::ChangeName { suggestions }
     } else if !possibilities.is_empty() {
         // Show the full list if all possible values for a given name, but don't do it
         // for names as the possibilities could be very long
@@ -343,7 +341,7 @@ pub(super) fn unexpected_cfg_value(
                 possibilities.clone(),
                 FilterWellKnownNames::No,
             );
-            lints::unexpected_cfg_value::ExpectedValues {
+            errors::unexpected_cfg_value::ExpectedValues {
                 name,
                 have_none_possibility,
                 possibilities: possibilities.into(),
@@ -354,7 +352,7 @@ pub(super) fn unexpected_cfg_value(
         let suggestion = if let Some((value, value_span)) = value {
             // Suggest the most probable if we found one
             if let Some(best_match) = find_best_match_for_name(&possibilities, value, None) {
-                Some(lints::unexpected_cfg_value::ChangeValueSuggestion::SimilarName {
+                Some(errors::unexpected_cfg_value::ChangeValueSuggestion::SimilarName {
                     span: value_span,
                     best_match,
                 })
@@ -362,7 +360,7 @@ pub(super) fn unexpected_cfg_value(
                 None
             }
         } else if let &[first_possibility] = &possibilities[..] {
-            Some(lints::unexpected_cfg_value::ChangeValueSuggestion::SpecifyValue {
+            Some(errors::unexpected_cfg_value::ChangeValueSuggestion::SpecifyValue {
                 span: name_span.shrink_to_hi(),
                 first_possibility,
             })
@@ -370,21 +368,21 @@ pub(super) fn unexpected_cfg_value(
             None
         };
 
-        lints::unexpected_cfg_value::CodeSuggestion::ChangeValue { expected_values, suggestion }
+        errors::unexpected_cfg_value::CodeSuggestion::ChangeValue { expected_values, suggestion }
     } else if have_none_possibility {
         let suggestion =
-            value.map(|(_value, value_span)| lints::unexpected_cfg_value::RemoveValueSuggestion {
+            value.map(|(_value, value_span)| errors::unexpected_cfg_value::RemoveValueSuggestion {
                 span: name_span.shrink_to_hi().to(value_span),
             });
-        lints::unexpected_cfg_value::CodeSuggestion::RemoveValue { suggestion, name }
+        errors::unexpected_cfg_value::CodeSuggestion::RemoveValue { suggestion, name }
     } else {
         let span = if let Some((_value, value_span)) = value {
             name_span.to(value_span)
         } else {
             name_span
         };
-        let suggestion = lints::unexpected_cfg_value::RemoveConditionSuggestion { span };
-        lints::unexpected_cfg_value::CodeSuggestion::RemoveCondition { suggestion, name }
+        let suggestion = errors::unexpected_cfg_value::RemoveConditionSuggestion { span };
+        errors::unexpected_cfg_value::CodeSuggestion::RemoveCondition { suggestion, name }
     };
 
     // We don't want to encourage people to add values to a well-known names, as these are
@@ -405,32 +403,32 @@ pub(super) fn unexpected_cfg_value(
     let invocation_help = if is_from_cargo {
         let help = if name == sym::feature && !is_from_external_macro {
             if let Some((value, _value_span)) = value {
-                Some(lints::unexpected_cfg_value::CargoHelp::AddFeature { value })
+                Some(errors::unexpected_cfg_value::CargoHelp::AddFeature { value })
             } else {
-                Some(lints::unexpected_cfg_value::CargoHelp::DefineFeatures)
+                Some(errors::unexpected_cfg_value::CargoHelp::DefineFeatures)
             }
         } else if can_suggest_adding_value && !is_from_external_macro {
-            Some(lints::unexpected_cfg_value::CargoHelp::Other(cargo_help_sub(sess, &inst)))
+            Some(errors::unexpected_cfg_value::CargoHelp::Other(cargo_help_sub(sess, &inst)))
         } else {
             None
         };
-        lints::unexpected_cfg_value::InvocationHelp::Cargo {
+        errors::unexpected_cfg_value::InvocationHelp::Cargo {
             help,
-            macro_help: cargo_macro_help(tcx, name_span),
+            macro_help: cargo_macro_help(sess_and_name.crate_name, name_span),
         }
     } else {
         let help = if can_suggest_adding_value {
-            Some(lints::UnexpectedCfgRustcHelp::new(&inst(EscapeQuotes::No)))
+            Some(errors::UnexpectedCfgRustcHelp::new(&inst(EscapeQuotes::No)))
         } else {
             None
         };
-        lints::unexpected_cfg_value::InvocationHelp::Rustc {
+        errors::unexpected_cfg_value::InvocationHelp::Rustc {
             help,
             macro_help: rustc_macro_help(name_span),
         }
     };
 
-    lints::UnexpectedCfgValue {
+    errors::UnexpectedCfgValue {
         code_sugg,
         invocation_help,
         has_value: value.is_some(),
