+++ compiler/rustc_attr_parsing/src/context.rs
@@ -11,7 +11,6 @@ use rustc_data_structures::sync::{DynSend, DynSync};
 use rustc_errors::{Diag, DiagCtxtHandle, Diagnostic, Level, MultiSpan};
 use rustc_feature::{AttrSuggestionStyle, AttributeTemplate};
 use rustc_hir::attrs::AttributeKind;
-use rustc_hir::lints::AttributeLintKind;
 use rustc_hir::{AttrPath, HirId};
 use rustc_parse::parser::Recovery;
 use rustc_session::Session;
@@ -459,18 +458,6 @@ pub struct AcceptContext<'f, 'sess, S: Stage> {
 impl<'f, 'sess: 'f, S: Stage> SharedContext<'f, 'sess, S> {
     pub(crate) fn emit_err(&self, diag: impl for<'x> Diagnostic<'x>) -> ErrorGuaranteed {
         self.stage.emit_err(&self.sess, diag)
-    }
-
-    /// Emit a lint. This method is somewhat special, since lints emitted during attribute parsing
-    /// must be delayed until after HIR is built. This method will take care of the details of
-    /// that.
-    pub(crate) fn emit_lint(
-        &mut self,
-        lint: &'static Lint,
-        kind: AttributeLintKind,
-        span: impl Into<MultiSpan>,
-    ) {
-        self.emit_lint_inner(lint, EmitAttribute::Static(kind), span);
     }
 
     /// Emit a lint. This method is somewhat special, since lints emitted during attribute parsing
