+++ compiler/rustc_lint/src/lints.rs
@@ -2653,378 +2653,6 @@ pub(crate) enum InvalidAsmLabel {
     },
 }
 
-#[derive(Subdiagnostic)]
-pub(crate) enum UnexpectedCfgCargoHelp {
-    #[help("consider using a Cargo feature instead")]
-    #[help(
-        "or consider adding in `Cargo.toml` the `check-cfg` lint config for the lint:{$cargo_toml_lint_cfg}"
-    )]
-    LintCfg { cargo_toml_lint_cfg: String },
-    #[help("consider using a Cargo feature instead")]
-    #[help(
-        "or consider adding in `Cargo.toml` the `check-cfg` lint config for the lint:{$cargo_toml_lint_cfg}"
-    )]
-    #[help("or consider adding `{$build_rs_println}` to the top of the `build.rs`")]
-    LintCfgAndBuildRs { cargo_toml_lint_cfg: String, build_rs_println: String },
-}
-
-impl UnexpectedCfgCargoHelp {
-    fn cargo_toml_lint_cfg(unescaped: &str) -> String {
-        format!(
-            "\n [lints.rust]\n unexpected_cfgs = {{ level = \"warn\", check-cfg = ['{unescaped}'] }}"
-        )
-    }
-
-    pub(crate) fn lint_cfg(unescaped: &str) -> Self {
-        UnexpectedCfgCargoHelp::LintCfg {
-            cargo_toml_lint_cfg: Self::cargo_toml_lint_cfg(unescaped),
-        }
-    }
-
-    pub(crate) fn lint_cfg_and_build_rs(unescaped: &str, escaped: &str) -> Self {
-        UnexpectedCfgCargoHelp::LintCfgAndBuildRs {
-            cargo_toml_lint_cfg: Self::cargo_toml_lint_cfg(unescaped),
-            build_rs_println: format!("println!(\"cargo::rustc-check-cfg={escaped}\");"),
-        }
-    }
-}
-
-#[derive(Subdiagnostic)]
-#[help("to expect this configuration use `{$cmdline_arg}`")]
-pub(crate) struct UnexpectedCfgRustcHelp {
-    pub cmdline_arg: String,
-}
-
-impl UnexpectedCfgRustcHelp {
-    pub(crate) fn new(unescaped: &str) -> Self {
-        Self { cmdline_arg: format!("--check-cfg={unescaped}") }
-    }
-}
-
-#[derive(Subdiagnostic)]
-#[note(
-    "using a cfg inside a {$macro_kind} will use the cfgs from the destination crate and not the ones from the defining crate"
-)]
-#[help("try referring to `{$macro_name}` crate for guidance on how handle this unexpected cfg")]
-pub(crate) struct UnexpectedCfgRustcMacroHelp {
-    pub macro_kind: &'static str,
-    pub macro_name: Symbol,
-}
-
-#[derive(Subdiagnostic)]
-#[note(
-    "using a cfg inside a {$macro_kind} will use the cfgs from the destination crate and not the ones from the defining crate"
-)]
-#[help("try referring to `{$macro_name}` crate for guidance on how handle this unexpected cfg")]
-#[help(
-    "the {$macro_kind} `{$macro_name}` may come from an old version of the `{$crate_name}` crate, try updating your dependency with `cargo update -p {$crate_name}`"
-)]
-pub(crate) struct UnexpectedCfgCargoMacroHelp {
-    pub macro_kind: &'static str,
-    pub macro_name: Symbol,
-    pub crate_name: Symbol,
-}
-
-#[derive(Diagnostic)]
-#[diag("unexpected `cfg` condition name: `{$name}`")]
-pub(crate) struct UnexpectedCfgName {
-    #[subdiagnostic]
-    pub code_sugg: unexpected_cfg_name::CodeSuggestion,
-    #[subdiagnostic]
-    pub invocation_help: unexpected_cfg_name::InvocationHelp,
-
-    pub name: Symbol,
-}
-
-pub(crate) mod unexpected_cfg_name {
-    use rustc_errors::DiagSymbolList;
-    use rustc_macros::Subdiagnostic;
-    use rustc_span::{Ident, Span, Symbol};
-
-    #[derive(Subdiagnostic)]
-    pub(crate) enum CodeSuggestion {
-        #[help("consider defining some features in `Cargo.toml`")]
-        DefineFeatures,
-        #[multipart_suggestion(
-            "there is a similar config predicate: `version(\"..\")`",
-            applicability = "machine-applicable"
-        )]
-        VersionSyntax {
-            #[suggestion_part(code = "(")]
-            between_name_and_value: Span,
-            #[suggestion_part(code = ")")]
-            after_value: Span,
-        },
-        #[suggestion(
-            "there is a config with a similar name and value",
-            applicability = "maybe-incorrect",
-            code = "{code}"
-        )]
-        SimilarNameAndValue {
-            #[primary_span]
-            span: Span,
-            code: String,
-        },
-        #[suggestion(
-            "there is a config with a similar name and no value",
-            applicability = "maybe-incorrect",
-            code = "{code}"
-        )]
-        SimilarNameNoValue {
-            #[primary_span]
-            span: Span,
-            code: String,
-        },
-        #[suggestion(
-            "there is a config with a similar name and different values",
-            applicability = "maybe-incorrect",
-            code = "{code}"
-        )]
-        SimilarNameDifferentValues {
-            #[primary_span]
-            span: Span,
-            code: String,
-            #[subdiagnostic]
-            expected: Option<ExpectedValues>,
-        },
-        #[suggestion(
-            "there is a config with a similar name",
-            applicability = "maybe-incorrect",
-            code = "{code}"
-        )]
-        SimilarName {
-            #[primary_span]
-            span: Span,
-            code: String,
-            #[subdiagnostic]
-            expected: Option<ExpectedValues>,
-        },
-        SimilarValues {
-            #[subdiagnostic]
-            with_similar_values: Vec<FoundWithSimilarValue>,
-            #[subdiagnostic]
-            expected_names: Option<ExpectedNames>,
-        },
-        #[suggestion(
-            "you may have meant to use `{$literal}` (notice the capitalization). Doing so makes this predicate evaluate to `{$literal}` unconditionally",
-            applicability = "machine-applicable",
-            style = "verbose",
-            code = "{literal}"
-        )]
-        BooleanLiteral {
-            #[primary_span]
-            span: Span,
-            literal: bool,
-        },
-    }
-
-    #[derive(Subdiagnostic)]
-    #[help("expected values for `{$best_match}` are: {$possibilities}")]
-    pub(crate) struct ExpectedValues {
-        pub best_match: Symbol,
-        pub possibilities: DiagSymbolList,
-    }
-
-    #[derive(Subdiagnostic)]
-    #[suggestion(
-        "found config with similar value",
-        applicability = "maybe-incorrect",
-        code = "{code}"
-    )]
-    pub(crate) struct FoundWithSimilarValue {
-        #[primary_span]
-        pub span: Span,
-        pub code: String,
-    }
-
-    #[derive(Subdiagnostic)]
-    #[help_once(
-        "expected names are: {$possibilities}{$and_more ->
-            [0] {\"\"}
-            *[other] {\" \"}and {$and_more} more
-        }"
-    )]
-    pub(crate) struct ExpectedNames {
-        pub possibilities: DiagSymbolList<Ident>,
-        pub and_more: usize,
-    }
-
-    #[derive(Subdiagnostic)]
-    pub(crate) enum InvocationHelp {
-        #[note(
-            "see <https://doc.rust-lang.org/nightly/rustc/check-cfg/cargo-specifics.html> for more information about checking conditional configuration"
-        )]
-        Cargo {
-            #[subdiagnostic]
-            macro_help: Option<super::UnexpectedCfgCargoMacroHelp>,
-            #[subdiagnostic]
-            help: Option<super::UnexpectedCfgCargoHelp>,
-        },
-        #[note(
-            "see <https://doc.rust-lang.org/nightly/rustc/check-cfg.html> for more information about checking conditional configuration"
-        )]
-        Rustc {
-            #[subdiagnostic]
-            macro_help: Option<super::UnexpectedCfgRustcMacroHelp>,
-            #[subdiagnostic]
-            help: super::UnexpectedCfgRustcHelp,
-        },
-    }
-}
-
-#[derive(Diagnostic)]
-#[diag(
-    "unexpected `cfg` condition value: {$has_value ->
-        [true] `{$value}`
-        *[false] (none)
-    }"
-)]
-pub(crate) struct UnexpectedCfgValue {
-    #[subdiagnostic]
-    pub code_sugg: unexpected_cfg_value::CodeSuggestion,
-    #[subdiagnostic]
-    pub invocation_help: unexpected_cfg_value::InvocationHelp,
-
-    pub has_value: bool,
-    pub value: String,
-}
-
-pub(crate) mod unexpected_cfg_value {
-    use rustc_errors::DiagSymbolList;
-    use rustc_macros::Subdiagnostic;
-    use rustc_span::{Span, Symbol};
-
-    #[derive(Subdiagnostic)]
-    pub(crate) enum CodeSuggestion {
-        ChangeValue {
-            #[subdiagnostic]
-            expected_values: ExpectedValues,
-            #[subdiagnostic]
-            suggestion: Option<ChangeValueSuggestion>,
-        },
-        #[note("no expected value for `{$name}`")]
-        RemoveValue {
-            #[subdiagnostic]
-            suggestion: Option<RemoveValueSuggestion>,
-
-            name: Symbol,
-        },
-        #[note("no expected values for `{$name}`")]
-        RemoveCondition {
-            #[subdiagnostic]
-            suggestion: RemoveConditionSuggestion,
-
-            name: Symbol,
-        },
-        ChangeName {
-            #[subdiagnostic]
-            suggestions: Vec<ChangeNameSuggestion>,
-        },
-    }
-
-    #[derive(Subdiagnostic)]
-    pub(crate) enum ChangeValueSuggestion {
-        #[suggestion(
-            "there is a expected value with a similar name",
-            code = r#""{best_match}""#,
-            applicability = "maybe-incorrect"
-        )]
-        SimilarName {
-            #[primary_span]
-            span: Span,
-            best_match: Symbol,
-        },
-        #[suggestion(
-            "specify a config value",
-            code = r#" = "{first_possibility}""#,
-            applicability = "maybe-incorrect"
-        )]
-        SpecifyValue {
-            #[primary_span]
-            span: Span,
-            first_possibility: Symbol,
-        },
-    }
-
-    #[derive(Subdiagnostic)]
-    #[suggestion("remove the value", code = "", applicability = "maybe-incorrect")]
-    pub(crate) struct RemoveValueSuggestion {
-        #[primary_span]
-        pub span: Span,
-    }
-
-    #[derive(Subdiagnostic)]
-    #[suggestion("remove the condition", code = "", applicability = "maybe-incorrect")]
-    pub(crate) struct RemoveConditionSuggestion {
-        #[primary_span]
-        pub span: Span,
-    }
-
-    #[derive(Subdiagnostic)]
-    #[note(
-        "expected values for `{$name}` are: {$have_none_possibility ->
-            [true] {\"(none), \"}
-            *[false] {\"\"}
-        }{$possibilities}{$and_more ->
-            [0] {\"\"}
-            *[other] {\" \"}and {$and_more} more
-        }"
-    )]
-    pub(crate) struct ExpectedValues {
-        pub name: Symbol,
-        pub have_none_possibility: bool,
-        pub possibilities: DiagSymbolList,
-        pub and_more: usize,
-    }
-
-    #[derive(Subdiagnostic)]
-    #[suggestion(
-        "`{$value}` is an expected value for `{$name}`",
-        code = "{name}",
-        applicability = "maybe-incorrect",
-        style = "verbose"
-    )]
-    pub(crate) struct ChangeNameSuggestion {
-        #[primary_span]
-        pub span: Span,
-        pub name: Symbol,
-        pub value: Symbol,
-    }
-
-    #[derive(Subdiagnostic)]
-    pub(crate) enum InvocationHelp {
-        #[note(
-            "see <https://doc.rust-lang.org/nightly/rustc/check-cfg/cargo-specifics.html> for more information about checking conditional configuration"
-        )]
-        Cargo {
-            #[subdiagnostic]
-            help: Option<CargoHelp>,
-            #[subdiagnostic]
-            macro_help: Option<super::UnexpectedCfgCargoMacroHelp>,
-        },
-        #[note(
-            "see <https://doc.rust-lang.org/nightly/rustc/check-cfg.html> for more information about checking conditional configuration"
-        )]
-        Rustc {
-            #[subdiagnostic]
-            help: Option<super::UnexpectedCfgRustcHelp>,
-            #[subdiagnostic]
-            macro_help: Option<super::UnexpectedCfgRustcMacroHelp>,
-        },
-    }
-
-    #[derive(Subdiagnostic)]
-    pub(crate) enum CargoHelp {
-        #[help("consider adding `{$value}` as a feature in `Cargo.toml`")]
-        AddFeature {
-            value: Symbol,
-        },
-        #[help("consider defining some features in `Cargo.toml`")]
-        DefineFeatures,
-        Other(#[subdiagnostic] super::UnexpectedCfgCargoHelp),
-    }
-}
-
 #[derive(Diagnostic)]
 #[diag("creating a {$shared_label}reference to mutable static")]
 pub(crate) struct RefOfMutStatic<'a> {
